requirements.txt
Appium-Python-Client==0.26
XlsxWriter==0.9.6
BeautifulRequests==0.2
PyYAML==3.12
XlsxWriter==0.9.6
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
gspread==3.0.0
httplib2==0.11.3
idna==2.7
more-itertools==4.2.0
multidict==4.3.1
oauth2client==4.1.2
pip==9.0.3
pluggy==0.6.0
py==1.5.3
pyasn1==0.4.3
pyasn1-modules==0.2.2
pytest==3.6.2
requests==2.19.1
rsa==3.4.2
selenium==3.12.0
setuptools==39.0.1
six==1.11.0
urllib3==1.23
google-api-python-client
apiclient